+++ packages/dom/src/platform/getDocumentElement.ts
@@ -1,7 +1,1 @@
-import {isNode} from '@floating-ui/utils/dom';
-
-export function getDocumentElement(node: Node | Window): HTMLElement {
-  return (
-    (isNode(node) ? node.ownerDocument : node.document) || window.document
-  )?.documentElement;
-}
+export {getDocumentElement} from '@floating-ui/utils/dom';
