+++ packages/dom/src/utils/getBoundingClientRect.ts
@@ -1,12 +1,11 @@
 import type {ClientRectObject, VirtualElement} from '@floating-ui/core';
 import {rectToClientRect} from '@floating-ui/core';
-import {getWindow} from '@floating-ui/utils/dom';
+import {createCoords} from '@floating-ui/utils';
+import {getWindow, unwrapElement} from '@floating-ui/utils/dom';
 
 import {getScale} from '../platform/getScale';
 import {isElement} from '../platform/isElement';
 import {getVisualOffsets, shouldAddVisualOffsets} from './getVisualOffsets';
-import {createCoords} from './math';
-import {unwrapElement} from './unwrapElement';
 
 export function getBoundingClientRect(
   element: Element | VirtualElement,
