--- packages/dom/src/utils/getComputedStyle.ts
@@ -1,5 +1,0 @@
-import {getWindow} from '@floating-ui/utils/dom';
-
-export function getComputedStyle(element: Element) {
-  return getWindow(element).getComputedStyle(element);
-}
