--- packages/dom/src/utils/getNearestOverflowAncestor.ts
@@ -1,23 +1,0 @@
-import {
-  isHTMLElement,
-  isLastTraversableNode,
-  isOverflowElement,
-} from '@floating-ui/utils/dom';
-
-import {getParentNode} from './getParentNode';
-
-export function getNearestOverflowAncestor(node: Node): HTMLElement {
-  const parentNode = getParentNode(node);
-
-  if (isLastTraversableNode(parentNode)) {
-    return node.ownerDocument
-      ? node.ownerDocument.body
-      : (node as Document).body;
-  }
-
-  if (isHTMLElement(parentNode) && isOverflowElement(parentNode)) {
-    return parentNode;
-  }
-
-  return getNearestOverflowAncestor(parentNode);
-}
