+++ packages/dom/src/utils/getWindowScrollBarX.ts
@@ -1,6 +1,7 @@
+import {getNodeScroll} from '@floating-ui/utils/dom';
+
 import {getDocumentElement} from '../platform/getDocumentElement';
 import {getBoundingClientRect} from './getBoundingClientRect';
-import {getNodeScroll} from './getNodeScroll';
 
 export function getWindowScrollBarX(element: Element): number {
   // If <html> has a CSS width greater than the viewport, then this will be
