--- packages/dom/src/utils/math.ts
@@ -1,5 +1,0 @@
-export const min = Math.min;
-export const max = Math.max;
-export const round = Math.round;
-export const floor = Math.floor;
-export const createCoords = (v: number) => ({x: v, y: v});
